+++ frontend/package.json
@@ -18,7 +18,6 @@
     "@types/react": "^17.0.38",
     "@types/react-dom": "^17.0.11",
     "bootstrap": "^5.1.3",
-    "chart.js": "^3.7.0",
     "jquery": "^3.6.0",
     "popper.js": "^1.16.1",
     "react": "^17.0.2",
